+++ src/App.js
@@ -1,11 +1,14 @@
-function App() {
+import React from 'react'
+import {Booksform} from './containers/BooksForm'
+import {Bookslist} from './containers/BooksList'
+
+const App = () => {
   return (
     <div className="App">
-      <header className="App-header">
-       <h1>React Bookstore</h1>
-      </header>
+      <Booksform />
+      <Bookslist />
     </div>
-  );
-};
+  )
+}
 
 export default App;
