+++ src/containers/BooksList.js
@@ -1,16 +1,19 @@
 
-const Bookslist = () => {
-   return (
-    <>
-      <table>
-        <tr>
-            <th>ID</th>
-            <th>Title</th>
-            <th>Category</th>
-        </tr>
-      </table>
-    </>
-   )
-}
+import { Provider } from "react-redux";
+const store = createStore(routerReducer)
 
-export default Bookslist
+ReactDOM.render (
+  <Provider store={store}>
+        <>
+          <table>
+            <tr>
+                <th>ID</th>
+                <th>Title</th>
+                <th>Category</th>
+            </tr>
+          </table>
+        </>
+  </Provider>
+);
+
+export default Bookslist;
